feat: ajustes de modulos

+++ src/modules/roles/hooks/useRolesCommand.ts
@@ -4,23 +4,33 @@ import { deleteRoleService } from "../services/deleteRole.service";
 import { RoleCreate, RoleUpdate } from "@/common/models/Role";
 import { createRoleService } from "../services/createRole.service";
 import { updateRoleService } from "../services/updateRole.service";
+import { useConfirmation } from "@/common/hooks/useConfirmation";
+import { toast } from "react-toastify";
 
 export const useRolesCommand = (refresh?: () => void) => {
    const { fetchData } = useFetch()
+   const confirmation = useConfirmation()
    const [loadingAction, setLoadingAction] = useState(false);
    const [errors, setErrors] = useState<{ [key: string]: string }>({})
 
    const deleteRole = async (id: string) => {
       try {
          setLoadingAction(true)
+         const confirmed = await confirmation({
+            title: "Eliminar rol",
+            message: "¿Estás seguro de que deseas eliminar este rol?",
+         });
+         if (!confirmed.isConfirmed) {
+            return;
+         }
          const response = await deleteRoleService(fetchData)(id)
          if (refresh) {
             refresh();
          }
-         alert("Rol eliminado correctamente")
-         return response
+         toast.success("Rol eliminado correctamente");
+         return response;
       } catch (error) {
-         alert(error.message)
+         toast.error(error.message);
       } finally {
          setLoadingAction(false)
       }
@@ -33,10 +43,10 @@ export const useRolesCommand = (refresh?: () => void) => {
          if (refresh) {
             refresh();
          }
-         alert("Rol creado correctamente")
-         return response
+         toast.success("Rol creado correctamente");
+         return response;
       } catch (error) {
-         alert(error.message)
+         toast.error(error.message);
       } finally {
          setLoadingAction(false)
       }
@@ -45,14 +55,21 @@ export const useRolesCommand = (refresh?: () => void) => {
    const updateRole = async (role: RoleUpdate) => {
       try {
          setLoadingAction(true)
+         const confirmed = await confirmation({
+            title: "Actualizar rol",
+            message: "¿Estás seguro de que deseas actualizar este rol?",
+         });
+         if (!confirmed.isConfirmed) {
+            return;
+         }
          const response = await updateRoleService(fetchData)(role)
          if (refresh) {
             refresh();
          }
-         alert("Rol actualizado correctamente")
+         toast.success("Rol actualizado correctamente");
          return response
       } catch (error) {
-         alert(error.message)
+         toast.error(error.message);
       } finally {
          setLoadingAction(false)
       }
